+++ Editor/Constants.cs
@@ -6,6 +6,7 @@ namespace SuperUnityBuild.BuildTool
         public const string BuildActionsDirectoryName = "BuildActions";
         public const string DefaultSettingsFileName = "SuperUnityBuildSettings";
         public const string EditorDirectoryName = "Editor";
+        public const string PackageName = "com.github.superunitybuild.buildtool";
         public const string RootDirectoryName = "SuperUnityBuild";
         public const string SettingsDirectoryName = "Settings";
     }