+++ Editor/Generic/TokensUtility.cs
@@ -1,5 +1,8 @@
 using System;
 using System.Text;
+using System.Text.RegularExpressions;
+using UnityEditor;
+using UnityEngine;
 
 namespace SuperUnityBuild.BuildTool
 {
@@ -51,6 +54,27 @@ namespace SuperUnityBuild.BuildTool
                 .Replace("$TIME", buildTime.ToString("hhmmss"));
         }
 
+        public static string ResolveBuildTimeUtilityTokens(string prototype, DateTime buildTime)
+        {
+            StringBuilder sb = new StringBuilder(prototype ?? "");
+
+            // Regex = (?:\$DAYSSINCE\(")([^"]*)(?:"\))
+            Match match = Regex.Match(prototype, "(?:\\$DAYSSINCE\\(\")([^\"]*)(?:\"\\))");
+            while (match.Success)
+            {
+                int daysSince = DateTime.TryParse(match.Groups[1].Value, out DateTime parsedTime) ?
+                    buildTime.Subtract(parsedTime).Days :
+                    0;
+
+                sb.Replace(match.Captures[0].Value, daysSince.ToString());
+                match = match.NextMatch();
+            }
+
+            sb.Replace("$SECONDS", (buildTime.TimeOfDay.TotalSeconds / 15f).ToString("F0"));
+
+            return sb.ToString();
+        }
+
         public static string ResolveBuildVersionToken(string prototype)
         {
             return (prototype ?? "")
@@ -59,10 +83,49 @@ namespace SuperUnityBuild.BuildTool
 
         public static string ResolveBuildVersionTokens(string prototype)
         {
-            prototype = prototype ?? "";
+            prototype ??= "";
 
             prototype = ResolveBuildVersionToken(prototype);
             prototype = ResolveBuildNumberToken(prototype);
+
+            return prototype;
+        }
+
+        public static string ResolveBuildWordTokens(string prototype)
+        {
+            prototype ??= "";
+
+            prototype = ReplaceTokenFromFile(prototype, "$NOUN", "nouns.txt");
+            prototype = ReplaceTokenFromFile(prototype, "$ADJECTIVE", "adjectives.txt");
+
+            return prototype;
+        }
+
+        private static string ReplaceTokenFromFile(string prototype, string token, string filename)
+        {
+            prototype ??= "";
+
+            if (prototype.IndexOf(token) > -1)
+            {
+                TextAsset textAsset = AssetDatabase.LoadAssetAtPath<TextAsset>($"Packages/{Constants.PackageName}/Editor/{filename}");
+
+                if (textAsset != null)
+                {
+                    string[] lines = textAsset.text.Split("\n");
+                    int index = prototype.IndexOf(token, 0);
+                    StringBuilder sb = new StringBuilder(prototype);
+
+                    while (index > -1)
+                    {
+                        string noun = lines[UnityEngine.Random.Range(0, lines.Length - 1)].ToUpper();
+
+                        sb.Replace(token, noun, index, token.Length);
+
+                        prototype = sb.ToString();
+                        index = prototype.IndexOf(token, index + 1);
+                    }
+                }
+            }
 
             return prototype;
         }